cassino/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pytube==15.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
yt-dlp==2025.8.22
